Fix buf parsing on CLIENT_SESSION_TRACK
Previously we continued buffer parsing if the CLIENT_SESSION_TRACK
capability flag was set, without first checking if the buffer was non
empty.

This logic assumes that with this capability flag set, that we always
get extra info back in the OK packet related to session state. This is
not the case, as evidenced by executing a single row insert, where the
OK packet returned will not have session state attached.

This causes us to parse an empty buffer, which modified earlier buffer
contents, resulting in incorrect result for `Affected Rows`. The result
this caused was always 0, which wasn't caught by current testing.

We replicated this with the following real world test case:

CREATE TABLE t1 (id INT);
INSERT INTO TABLE t1 (id) VALUES (4);
rows_affected: 0

Here's the Ok packet that came back from this insert:

"x00/x01/x00/x02/x00/x00/x00"

We can see that the rows affected should be 1, and we also see that no
extra session info was attached to this packet. If we log at the
beginning of parsing, right after we parse the buffer for the
rows_affected data, we get the correct result. If we instead put a log
at the end of the method right before the Ok body is returned, we get a
value of 0, indicating that parsing an empty buffer caused us to modify
earlier buffer results.

If we instead do a multi-row insert, we get the correct result:

CREATE TABLE t1 (id INT);
INSERT INTO TABLE t1 (id) VALUES (1), (2), (3);
rows_affected: 3

This is because session state is passed back in the case of multi-row
inserts.

This commit fixes this by checking that the buffer is non empty in the
case that the CLIENT_SESSION_TRACK compatibility flag is set. This also
updates our testing so our basic OK packet example has a rows affected
result that is non-zero.

This has also been tested externally with a demo application using
`mysql_async` to confirm that rows_affected is correct in the single row
insert case with this capability flag set.