Optional Metadata fix for Minimal Row Based Replication
This commit fixes the issue where we only advance the iterators for
columns that are present in the row image. The problem is that in
the table map event, the optional metadata is populated for all
columns, thus we need to advance the iterators even for columns that
are not present in the row event, otherwise we will be reading
metadata for the wrong column.

The fix is to keep all the logic of advancing the iterators for all
columns, and only populate the Column and value objects for the
columns that are present in the row event.

Fixes: #162